time_columns added

+++ delete_alert_rule.py
@@ -1,8 +1,23 @@
 from tool.helper import *
 
+"""
+This script is used to delete an alert rule from Grafana based on user requested UID.
+    - Print `exit` to exit without deleting.
+"""
+
+# Fetch all existing alert rules' uids
 result = client.get_all_alert_rules()
 print(f" >> Existing alert rules uids: {result} \n")
 
-target_alert_rule_uid = input(" >> Enter the UID of the alert rule to be deleted: ")
+# Ask user to input the UID of the alert rule to be deleted
+while True:
+    target_alert_rule_uid = input(" >> Enter the UID of the alert rule to be deleted: ")
 
-client.delete_alert_rule(target_alert_rule_uid)
+    if target_alert_rule_uid in result:
+        client.delete_alert_rule(target_alert_rule_uid)
+        break
+    elif target_alert_rule_uid.lower() == "exit":
+        print("   >> Exiting without deleting.")
+        break
+    else:
+        print(f" >> Alert rule with UID '{target_alert_rule_uid}' does not exist. Please try again.\n")
